README.MD
Don't let the name fool you! Yes there was definitely some humor involved, but the site was made to be a safe and anonymous place to get information on a topic that some people find intimidating or embarrassing. The app, as of now, allows you to post, read other user posts, edit/delete your post, and comment on all posts. 

This app was created using HTML/EJS, Mongoose/Mongo, Express, and Node. There are several additional features I would like to implement such as user sign in, allowing only user to edit their posts and delete comments if they wish. I'd also put glyphicons to show likes/dislikes for posts and a time/date stamp for when posts were made. My biggest challenge will be finding a way to categorize posts into either weeks or topic. As of now, there are suggested topic tags to copy and paste into the a reserved section of your post. This shows readers what your post will be about. However, I need to find a way to better organize once more and more posts are made. There is still much CSS to do to better the layout and form's appearance.

A big problem I had, and am still having, is with the comments show page. I don't really see a need for this page, nor a comment index page as it doesn't have a very useful functionality, so I will probably end up scratching them.

This is my first full stack app and was a lot of fun to build. Quite a bit of trial and error, but I learned a lot and look forward to adding to it.

A readme.md file with explanations of the technologies used, the approach was taken, unsolved problems, user stories, and notes to yourself so you can come back to your project later in the course and be able to pick up your train of thought, etc